refactor: tsup config
