+++ docs/next.config.js
@@ -40,6 +40,9 @@ const config = {
 }
 
 if (process.env.NODE_ENV === 'development') {
+  // We don't use Preconstruct anymore, but the Next.js plugin works great for
+  // development with our custom set up in packages' `tsup.config.ts`.
+  // https://github.com/preconstruct/preconstruct/tree/main/packages/next
   const withPreconstruct = require('@preconstruct/next')
   module.exports = withPreconstruct(withNextra(config))
 } else {
