+++ examples/_dev/src/components/Account.tsx
@@ -47,7 +47,7 @@ export const Account = () => {
         )}
       </div>
 
-      {false && (
+      {true && (
         <>
           {false && (
             <>
